requirements.txt
libusb
hidapi
streamlit
matplotlib